+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - `[jest-jasmine2]` Stop adding `:` after an error that has no message ([#9990](https://github.com/facebook/jest/pull/9990))
+- `[jest-diff]` Control no diff message color with `commonColor` in diff options ([#9997](https://github.com/facebook/jest/pull/9997))
 
 ### Chore & Maintenance